config: Move config init code
Move the config initialization code from main.go to
internal/config/config.go.

Suggested-by: Bruno Meneguele <bmeneg@redhat.com>
