+++ internal/config/config.go
@@ -3,16 +3,21 @@ package config
 import (
 	"bufio"
 	"bytes"
+	"crypto/tls"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"log"
+	"net/http"
 	"net/url"
 	"os"
+	"path"
 	"strings"
 	"syscall"
 
 	"github.com/spf13/viper"
+	gitlab "github.com/xanzy/go-gitlab"
+	"github.com/zaquestion/lab/internal/git"
 	"golang.org/x/crypto/ssh/terminal"
 )
 
@@ -141,4 +146,100 @@ func ConvertHCLtoTOML(oldpath string, newpath string, file string) {
 	// END HACK
 
 	fmt.Println("INFO: Converted old config", oldconfig, "to new config", newconfig)
+}
+
+func getUser(host, token string, skipVerify bool) string {
+	httpClient := &http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{
+				InsecureSkipVerify: skipVerify,
+			},
+		},
+	}
+	lab, _ := gitlab.NewClient(token, gitlab.WithHTTPClient(httpClient), gitlab.WithBaseURL(host+"/api/v4"))
+	u, _, err := lab.Users.CurrentUser()
+	if err != nil {
+		log.Fatal(err)
+	}
+	return u.Username
+}
+
+// LoadConfig() loads the main config file
+func LoadConfig() (string, string, string, bool) {
+
+	// Attempt to auto-configure for GitLab CI.
+	// Always do this before reading in the config file o/w CI will end up
+	// with the wrong data.
+	host, user, token := CI()
+	if host != "" && user != "" && token != "" {
+		return host, user, token, false
+	}
+
+	// Try to find XDG_CONFIG_HOME which is declared in XDG base directory
+	// specification and use it's location as the config directory
+	home, err := os.UserHomeDir()
+	if err != nil {
+		log.Fatal(err)
+	}
+	confpath := os.Getenv("XDG_CONFIG_HOME")
+	if confpath == "" {
+		confpath = path.Join(home, ".config")
+	}
+	labconfpath := confpath + "/lab"
+	if _, err := os.Stat(labconfpath); os.IsNotExist(err) {
+		os.MkdirAll(labconfpath, 0700)
+	}
+
+	// Convert old hcl files to toml format.
+	// NO NEW FILES SHOULD BE ADDED BELOW.
+	ConvertHCLtoTOML(".", ".", "lab")
+	ConvertHCLtoTOML(confpath, labconfpath, "lab")
+	var labgitDir string
+	gitDir, err := git.GitDir()
+	if err == nil {
+		labgitDir = gitDir + "/lab"
+		ConvertHCLtoTOML(gitDir, labgitDir, "lab")
+		ConvertHCLtoTOML(labgitDir, labgitDir, "show_metadata")
+	}
+
+	viper.SetConfigName("lab")
+	viper.SetConfigType("toml")
+	viper.AddConfigPath(".")
+	viper.AddConfigPath(labconfpath)
+	if labgitDir != "" {
+		viper.AddConfigPath(labgitDir)
+	}
+
+	viper.SetEnvPrefix("LAB")
+	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
+	viper.AutomaticEnv()
+
+	if _, ok := viper.ReadInConfig().(viper.ConfigFileNotFoundError); ok {
+		err := New(path.Join(labconfpath, "lab.toml"), os.Stdin)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		err = viper.ReadInConfig()
+		if err != nil {
+			log.Fatal(err)
+		}
+	}
+
+	host = viper.GetString("core.host")
+	user = viper.GetString("core.user")
+	token = viper.GetString("core.token")
+	tlsSkipVerify := viper.GetBool("tls.skip_verify")
+
+	if host != "" && user != "" && token != "" {
+		return host, user, token, tlsSkipVerify
+	}
+
+	user = getUser(host, token, tlsSkipVerify)
+	if strings.TrimSpace(os.Getenv("LAB_CORE_TOKEN")) == "" && strings.TrimSpace(os.Getenv("LAB_CORE_HOST")) == "" {
+		viper.Set("core.user", user)
+		viper.WriteConfig()
+	}
+
+	return host, user, token, tlsSkipVerify
 }
